backend/requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
nvidia-riva-client>=2.18.0
supabase>=2.19.0,<3.0.0
pydantic-settings>=2.10.1,<3.0.0
grpcio>=1.76.0,<2.0.0
